changed oldAngle variable in timer1 interrupt from being local variable to being a variable contained in encoder object, in order to fix bug with encoder.setHome() function

+++ src/uStepper.h
@@ -375,7 +375,7 @@ class uStepperEncoder
 public:
 	volatile int32_t angleMoved;			/**< Variable used to store that measured angle moved from the reference position */
 	uint16_t encoderOffset;				/**< Angle of the shaft at the reference position. */
-	volatile float oldAngle;			/**< Used to stored the previous measured angle for the speed measurement, and the calculation of angle moved from reference position */
+	volatile uint16_t oldAngle;			/**< Used to stored the previous measured angle for the speed measurement, and the calculation of angle moved from reference position */
 	volatile uint16_t angle;
 	volatile float curSpeed;			/**< Variable used to store the last measured rotational speed of the motor shaft */ 	
 	/**
